+++ notification/src/main/java/com/back/notification/app/strategy/SettlementCompletedStrategy.java
@@ -1,6 +1,7 @@
-package com.back.notification.domain;
+package com.back.notification.app.strategy;
 
 import com.back.common.Settlement.event.SettlementCompletedEvent;
+import com.back.notification.domain.Notification;
 import com.back.notification.domain.enums.Type;
 import com.back.notification.mapper.NotificationMapper;
 import lombok.RequiredArgsConstructor;
@@ -10,7 +11,7 @@ import java.util.List;
 
 @Component
 @RequiredArgsConstructor
-public class SettlementCompletedStrategy implements NotificationStrategy<SettlementCompletedEvent>{
+public class SettlementCompletedStrategy implements NotificationStrategy<SettlementCompletedEvent> {
     private final NotificationMapper mapper;
 
     public Type type(){
